fix: Move events endpoint to it's own path and router
