+++ meetupselector/api/routes/__init__.py
@@ -1,3 +1,4 @@
+from .events import router as events_router  # noqa: F401
 from .proposals import router as proposals_router  # noqa: F401
 from .topics import router as topics_router  # noqa: F401
 from .users import router as users_router  # noqa: F401
